backend/requirements.txt
flask
flask-cors
google-cloud-speech
google-cloud-storage
google-generativeai
werkzeug
